Extract validator
