+++ src/main/java/baseball/Controller/GameController.java
@@ -37,6 +37,7 @@ public class GameController {
         System.out.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
         System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
         String userReply = Console.readLine();
+        Validator.checkValidOfReply(userReply);
         return userReply.equals("1");
     }
 }
